+++ ktorm-core/src/main/kotlin/me/liuwj/ktorm/schema/Reflects.kt
@@ -1,9 +1,17 @@
 package me.liuwj.ktorm.schema
 
+import me.liuwj.ktorm.entity.Entity
+import java.lang.reflect.Method
 import java.lang.reflect.ParameterizedType
 import java.lang.reflect.Type
-import kotlin.reflect.KClass
-import kotlin.reflect.KType
+import java.util.*
+import kotlin.reflect.*
+import kotlin.reflect.full.createInstance
+import kotlin.reflect.full.declaredMemberProperties
+import kotlin.reflect.full.isSubclassOf
+import kotlin.reflect.jvm.javaGetter
+import kotlin.reflect.jvm.javaSetter
+import kotlin.reflect.jvm.jvmName
 
 /**
  * Base class used for obtaining full generic type information by subclassing.
@@ -80,4 +88,56 @@ inline fun <reified T> typeOf(): Type {
 @Deprecated("Do not use this function until the bug KT-28616 fixed", ReplaceWith("typeOf<T>()"))
 inline fun <reified T> kotlinTypeOf(): KType {
     return typeRef<T>().referencedKotlinType
+}
+
+internal val Method.kotlinProperty: Pair<KProperty1<*, *>, Boolean>? get() {
+    for (prop in declaringClass.kotlin.declaredMemberProperties) {
+        if (prop.javaGetter == this) {
+            return prop to true
+        }
+        if (prop is KMutableProperty<*> && prop.javaSetter == this) {
+            return prop to false
+        }
+    }
+    return null
+}
+
+internal val KClass<*>.defaultValue: Any get() {
+    val value: Any = try {
+        when {
+            this == Boolean::class -> false
+            this == Char::class -> 0.toChar()
+            this == Byte::class -> 0.toByte()
+            this == Short::class -> 0.toShort()
+            this == Int::class -> 0
+            this == Long::class -> 0L
+            this == Float::class -> 0.0F
+            this == Double::class -> 0.0
+            this == String::class -> ""
+            this.isSubclassOf(Entity::class) -> Entity.create(this)
+            this.java.isEnum -> this.java.enumConstants[0]
+            this.java.isArray -> this.java.componentType.createArray(0)
+            this == Set::class || this == MutableSet::class -> LinkedHashSet<Any?>()
+            this == List::class || this == MutableList::class -> ArrayList<Any?>()
+            this == Collection::class || this == MutableCollection::class -> ArrayList<Any?>()
+            this == Map::class || this == MutableMap::class -> LinkedHashMap<Any?, Any?>()
+            this == Queue::class || this == Deque::class -> LinkedList<Any?>()
+            this == SortedSet::class || this == NavigableSet::class -> TreeSet<Any?>()
+            this == SortedMap::class || this == NavigableMap::class -> TreeMap<Any?, Any?>()
+            else -> this.createInstance()
+        }
+    } catch (e: Throwable) {
+        throw IllegalArgumentException("Error creating a default value for non-null type: ${this.jvmName}", e)
+    }
+
+    if (this.isInstance(value)) {
+        return value
+    } else {
+        // never happens...
+        throw AssertionError("$value must be instance of $this")
+    }
+}
+
+private fun Class<*>.createArray(length: Int): Any {
+    return java.lang.reflect.Array.newInstance(this, length)
 }
