+++ ktorm-jackson/src/main/kotlin/me/liuwj/ktorm/jackson/JacksonExtensions.kt
@@ -2,7 +2,6 @@ package me.liuwj.ktorm.jackson
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.ObjectCodec
-import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.databind.cfg.MapperConfig
